some experiments on how to express the interpreter
interp.ml provides a fairly straight forward implementation
of interp.c in ocaml.  The intention is for this to dissapear.

interp2.ml is how I actually want to do it.  This expresses
the interpreter in a monad which is abtracted over a state type.
The instructions can then be evaluated directly, or run to
produce an AST.  We will then compile the hardware
statemachine from this description rather than directly writing
a massive statemachine.  This will allow us to experiment with
different architectures with the tradeoff that we will need to
do some analysis on the AST instead.  If it works.

Need to tidy up the AST representation a bit still.  The main
unresolved issue is how to express conditionals and loops.