data64 <-> repr64
Provide functions to convert from Obj.t to an array of int64
and back again.

With this we can (to some extent anyway) use the host ocaml
runtime to format and parse objects in the simulated
system simplifying some of the c_runtime code.

Use with care goes without saying...