Ocaml REPL runs up to caml_parse_engine
...and we're stuck.

caml_parse_engine is passed function values and it's tough to
figure out how these can pass from the simulated system into the
host environment.

what we need is an implemention in ocaml against the simulated
systems memory model (along the lines of parser.js in js_of_ocaml).

Another option is an ocaml version and modified compiler.

Or we bite the bullet and accept that things with very complex
c-bindings like this would be better served by properly
porting the runtime system and running on a real system.

Either way, we have now tested a pretty significant amount of
bytecode, can run a fair amount of real code (even if the
REPL is off limits for now) so perhaps its time to look at generating
the hardware.