Compile instructions into a single statemachine.
There are a few instructions not included yet (STOP, RAISE, DIV, MOD etc)
and some that can't work as written (C_CALL etc) but this is the basic
framework for bringing up a hardware simulation of a bytecode program.

There is no doubt this is an insane architecture at the moment.  I doubt
there are many FPGAs which would fit it.  Still, it allows us to start
to consider the system level architecture, get the compilation code
working and then start to consider how to optmisise it.